Improve api configuration classification
